Moved login exceptions to util.
